+++ libs/ledgerjs/packages/devices/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/devices
 
+## 8.0.5
+
+### Patch Changes
+
+- Updated dependencies [[`0f4293e9bf`](https://github.com/LedgerHQ/ledger-live/commit/0f4293e9bf9cac4c2a195efeb0831aab3d51933d)]:
+  - @ledgerhq/errors@6.13.0
+
+## 8.0.5-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`0f4293e9bf`](https://github.com/LedgerHQ/ledger-live/commit/0f4293e9bf9cac4c2a195efeb0831aab3d51933d)]:
+  - @ledgerhq/errors@6.13.0-next.0
+
 ## 8.0.4
 
 ### Patch Changes